requirements-dev.txt
# Used for example docs and/or docs guide
ipykernel
jupyter-sphinx
numpy
matplotlib
sphinx-design
sphinx-autodoc-typehints
sphinx-copybutton
nbsphinx==0.9.*

# Dev tooling
black~=23.3
isort~=5.11
mypy~=1.3.0
pytest==7.*

# Types
types-docutils
types-Pygments
